chore: creats new constants file, moves breadcrumbs logic into header component
